Replace useless shared pointers with unique pointers
This also means a whole bunch of refactoring and changing how the code
works, because now we explicitly track ownership.
- To accomodate this, the way Targets work has been changed.
- Member pointers in the AST have been made const, because once the AST is built
it shouldn't change. So the way data flows now, is that unique_ptrs are
built during the getExpression and exprpParse phases, and then they are
progressively moved into their final place during the exprParse phase,
and nothing is being moved around afterwards.